Hybrid Gen AI Telegram Bot/requirements.txt
python-telegram-bot==21.10
sentence-transformers==3.4.1
transformers==4.49.0
torch>=2.2.0
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
numpy==2.2.3
